requirements.txt
numpy-stl>=3.0.0
numpy>=1.21.0
customtkinter>=5.2.0
tkinterdnd2>=0.3.0


